% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibl_new.R
\name{ibl_new}
\alias{ibl_new}
\title{Iterates through the IBL Model}
\usage{
ibl_new(noptions, prevalues, ntrials, noise, decay, gambles, foregone)
}
\arguments{
\item{noptions}{The number of options available.}

\item{prevalues}{The number of simulated draws before the start of the
experiment.}

\item{ntrials}{The number of trials in the experiment.}

\item{noise}{The noise parameter of IBL.}

\item{decay}{The decoy parameter of IBL.}

\item{gambles}{A list of gambles containing payoffs in column 1 and
associated probabilities in column 2.}

\item{foregone}{Logical. Whether foregone outcomes are observed by the
decision maker.}
}
\value{
A column vector representing the participant's choice in each trial.
  The length of the vector is equal to the number of trials.
}
\description{
\code{ibl_new} is called by \code{ibl_simulation} and simulates one
participant going through the experiment.
}

