% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibl_simulation.R
\name{ibl_simulation}
\alias{ibl_simulation}
\title{Simulates the IBL Model}
\usage{
ibl_simulation(gambles, ntrials, prevalues, noise = 0.25, decay = 0.5,
  foregone = F, parallel = T, cores = NULL, iterations = 1000)
}
\arguments{
\item{gambles}{A list of gambles containing payoffs in column 1 and associated probabilities in column 2}

\item{ntrials}{The number of trials in the experiment}

\item{prevalues}{The number of simulated draws before the start of the experiment}

\item{noise}{The noise parameter of IBL}

\item{decay}{The decoy parameter of IBL}

\item{foregone}{Logical. Whether foregone outcomes are observed by the decision maker}

\item{parallel}{Logical. Should the model run using parallel processing?}

\item{cores}{Number of cores. If left blank, uses half of all available cores (fastest)}

\item{iterations}{The number of iterations for the simulation}
}
\value{
A matrix with the number of rows equal to \code{ntrials} and the number of columns equal to \code{iterations}.
}
\description{
Simulates the IBL Model
}
\examples{
Simple example needed. (After fixing how gambles are handled?)
}

